+++ core/ajax/mczremote.ajax.php
@@ -26,11 +26,16 @@ try {
     
     ajax::init();
 
+    if (init('action') == 'installTemplate') {
+        mczremote::installTemplate();
+        ajax::success();
+    }
+
     if (init('action') == 'health') {
         $data = md5(rand());
         log::add('mczremote', 'debug', 'health:: data:' . var_export($data, true));
-	$params = array('method' => 'health', 'data' => $data );
-	mczremote::daemon_send($params);
+        $params = array('method' => 'health', 'data' => $data );
+        mczremote::daemon_send($params);
         ajax::success();
     }