+++ plugin_info/configuration.php
@@ -85,6 +85,12 @@ if (!isConnect()) {
                 <input class="configKey form-control" data-l1key="TopicSub" />
             </div>
         </div>
+         <div class="form-group">
+          <label class="col-md-4 control-label">{{Installer template dans jMQTT}}</label>
+          <div class="col-md-4">
+            <a class="btn btn-warning" id="bt_InstallTemplate">{{Installer}}</a>
+          </div>
+        </div>
         <legend><i class="fas fa-university"></i> {{Démon}}</legend>
         <div class="form-group">
             <label class="col-sm-4 control-label">{{Port socket interne}} <sup><i class="fas fa-question-circle" title="{{Si le numéro de port est en conflit avec un autre service, mettez à jour ce champ en indiquant un numéro de port non utilisé par le système.}}"></i></sup></label>
@@ -94,4 +100,36 @@ if (!isConnect()) {
         </div>
   </fieldset>
 </form>
+<script>
+  $('#bt_InstallTemplate').off('click').on('click', function() {
+    $.ajax({
+      type: "POST",
+      url: "plugins/mczremote/core/ajax/mczremote.ajax.php",
+      data: {
+        action: "installTemplate"
+      },
+      dataType: 'json',
+      error: function(request, status, error) {
+        handleAjaxError(request, status, error);
+      },
+      success: function(data) {
+        if (data.state != 'ok') {
+          $('#div_alert').showAlert({
+            message: data.result,
+            level: 'danger'
+          });
+          return;
+        } else {
+          window.toastr.clear()
+          $('.pluginDisplayCard[data-plugin_id=' + $('#span_plugin_id').text() + ']').click()
+          $('#div_alert').showAlert({
+            message: '{{Installation réussie}}',
+            level: 'success'
+          });
+
+        }
+      }
+    });
+  });
+</script>