+++ docs/fr_FR/changelog.md
@@ -4,7 +4,9 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte
 
-
+# 06/02/2022
+- Copie du template MCZRemote vers jMQTT
+- Correction documentation
 
 # 19/01/2022
 - encrypte les infos de configuration dans la BDD